refactor(search-view): extract results list, command card, and content renderer
